refactor: remove DAO layer
