+++ src/main/java/icu/random/client/rest/RestClient.java
@@ -1,7 +1,6 @@
 package icu.random.client.rest;
 
 import kong.unirest.GetRequest;
-import kong.unirest.HttpResponse;
 
 /**
  * Rest client for Lipsum.com API
@@ -12,9 +11,7 @@ public interface RestClient {
   /**
    * Send HTTP GET request to Lipsum.com API
    *
-   * @return {@link HttpResponse} object
+   * @return {@link GetRequest} object
    */
-  <T> HttpResponse<T> get(String path, Class<T> clazz);
-
   GetRequest get(String path);
 }
